program that prints the first 50 Fibonacci numbers, starting with 1 and 2, followed by a new line

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	int a = 1, b = 2, c, run = 2;
+	long a = 1, b = 2, c, run = 2;
 
 	printf("%d, %d", a, b);
 	while (run < 50)
